+++ src/core/enums/IterableEventName.ts
@@ -23,4 +23,8 @@ export enum IterableEventName {
   handleEmbeddedMessageUpdateCalled = 'handleEmbeddedMessageUpdateCalled',
   /** Event that fires when embedded messaging is disabled */
   handleEmbeddedMessagingDisabledCalled = 'handleEmbeddedMessagingDisabledCalled',
+  /**
+   * Event that fires when Android keychain decryption fails (Android only).
+   */
+  handleDecryptionFailureCalled = 'handleDecryptionFailureCalled',
 }